+++ anchorpoint/__init__.py
@@ -1,4 +1,4 @@
 from anchorpoint.textselectors import TextQuoteSelector, TextPositionSelector
-from anchorpoint.textselectors import TextPositionSet
+from anchorpoint.textselectors import TextPositionSet, TextPositionSetFactory
 
 __version__ = "0.5.3"