+++ changelog.rst
@@ -8,7 +8,7 @@ dev
 - TextPositionSelector no longer has real_start and real_end that can differ from start and end
 - Selectors and TextPositionSets are Pydantic models
 - TextSelector is Pydantic model for either Quote or Position Selector
-- split Marshmallow SelectorSchema into Quote and Position schemas
+- remove Marshmallow schemas
 - update type annotations for TextPositionSelector.from_range
 - add TestQuoteSelector.as_unique_position method
 - TextPositionSet can include TextQuoteSelectors